Move LogSettings to a different package.
