+++ blueflood-dropwizard/src/main/java/com/rackspacecloud/blueflood/dw/logging/LogSettings.java
@@ -1,10 +1,8 @@
-package com.rackspacecloud.blueflood.dw;
+package com.rackspacecloud.blueflood.dw.logging;
 
 import java.io.IOException;
 import java.net.URL;
-import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.Properties;
